optimizer wmt14 dataset
